+++ aio/aio-proxy/aio_proxy/routes.py
@@ -4,9 +4,9 @@ import os
 import elasticsearch
 import sentry_sdk
 from aio_proxy.helper import serialize_error_text
-from aio_proxy.index import Siren
+from aio_proxy.search.index import Siren
 from aio_proxy.parameters import extract_parameters
-from aio_proxy.search_functions import search_es
+from aio_proxy.search.search_functions import search_es
 from aiohttp import web
 from dotenv import load_dotenv
 from elasticsearch_dsl import connections
